Add SessionProvider
