+++ services/orchest-webserver/client/src/lib/orchest/index.js
@@ -1,2 +1,4 @@
-export * from "./context";
-export * from "./provider";
+// @ts-check
+
+export { OrchestContext, useOrchest } from "./context";
+export { OrchestProvider } from "./provider";
